LU-3233 mdt: validate open handle cookies
Add a const void *h_owner member to struct portals_handle. Add a const
void *owner parameter to class_handle2object() which must be matched
by the h_owner member of the handle in addition to the cookie.  Adjust
the callers of class_handle2object() accordingly, using NULL as the
argument to the owner parameter, except in the case of
mdt_handle2mfd() where we add a mdt_export_data parameter which we use
as the owner when searching for a MFD. When allocating a new MFD, pass
a pointer to the mdt_export_data into mdt_mfd_new() and store it in
h_owner. This allows the MDT to validate that the client has not sent
the wrong open handle cookie, or sent the right cookie to the wrong
MDT.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I4c26504d808db0c27cfc2f8ebcf7e1fa19719170