LU-2919 hsm: Implementation of exclusive open
In this patch, a framework of lease is implemented. However,
only exclusive lease is supported right now.

To apply a lease, MDS_OPEN_LEASE must be set to open the file, EX mode
open lock is returned to the client side to hold a lease. From that
time on, if this file is opened again by other processes, the open
lock will be revoked so the client who holds the lease will know the
lease is already broken by checking that open lock.

To release a lease, normal close is used. The client will revoke the
open lock before sending CLOSE request.

The function ll_lease_open() is added for in kernel code to acquire a
lease on an inode. This may be extended to allow userspace to
manipulate leases via ioctls. Note that these leases are not to be
confused with the leases provided by the fcntl() flags F_SETLEASE and
F_GETLEASE and that this implementation does not aim to meet their
requirements.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I181edbb0b31fcf60858656170528d2ffda9c9bb6