LU-3587 changelogs: Account for changelog_ext_rec in CR_MAXSIZE
CR_MAXSIZE needs to account for an llog_changelog_rec that actually
contains a changelog_ext_rec structure rather than a changelog_rec.
With out doing so, a file size approaching the Linux kernel NAME_MAX
length that is renamed to a size also close to, or at, NAME_MAX will
exceed CR_MAXSIZE and trip an assertion.

Change-Id: I2d86955efcbd6d0fd28c87026a7720311b4f5598
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>