LU-3613 llapi: add call to get parent fids from link ea.
To get all <parent_fid>/name for an entry, it was necessary
to run path2fid on it, then call fid2path with the returned fid,
and finally call path2fid for each of the parent directories.
This patch simplifies this process and makes it lighter
by reading directly the value in link ea.
Added -p option to 'lfs path2fid' to call this function.
Added sanity 164.

Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Change-Id: I10f9316da605050be8fc709fe88d87abe86e22a4