LU-3539 protocol: Change UPDATE_OBJ RPC format
This patch updates struct update for DNE phase II.  See the DNE phase
II HLD for more information.

The comment describing the UPDATE_OBJ RPC format is already a bit out
of date.  This patch simplies and updates it so that the overall
structure is easier to see while the details are commented in-line.

The related type names are modified to be more consistent with the
terminology used in the HLD and the comment.

Change-Id: I4355b4c863abb448713115f17b80402ec96009f8
Signed-off-by: Li Wei <wei.g.li@intel.com>