LU-3539 osp: Fix a series of UPDATE_OBJ endianness bugs
Current UPDATE_OBJ code have the following problems:

  - Fields in "struct update_buf", for ub_bufs, were sent in
    senders' host endianness but expected to be little-endian by
    receivers.

  - Fields in "struct update", except for u_bufs, were sent in
    little endianness but used without swabbing by receivers.

  - Most u_bufs were sent in little endianness, but some were in
    senders' host endianness.

  - Sizes for a couple of string buffers in osp_md_declare_xattr_set()
    and osp_md_xattr_get() are off-by-one.

  - Error numbers in UPDATE_OBJ replies are not translated on both
    sides.

This patch fixes them by sending all data in UPDATE_OBJ requests and
replies in sender's host endianness following Lustre's usual practice.
It is up to the receiver to swab to its host endianness if needed.

Add update_reply_set_swabbed() and update_reply_need_swab() to allow
tracking whether a reply buffer has been swabbed already by the user.
Once swabbing has started, the reply buffer should always be marked
swabbed if returning an error, so it is not swabbed repeatedly if it
is accessed multiple times.  The UPDATE_FL_SWABBED flag is temporarily
stored in the ur_lens[] array, but should be moved over to our_datalen
when that is available.

Change-Id: Id3ffb248af181b4beff2b62c9cb46fc6d93ebbe5
Signed-off-by: Li Wei <wei.g.li@intel.com>
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>