LU-3485 tests: Add resize2fs testing for OST/MDT filesystems
Add testing of the MST and OST filesystems with resize2fs.
Currently, only offline resizing is possible until LU-137
is finished. Thus, each file system is unmounted and
e2fsck is run before and after the resize to check for
errors.

We are restricting the number of OSTs since large-lun
test 2 run time is based on the number of OSTs and can
time out if there are 16 OSTs as in the last large-lun
test results.

Test-Parameters: testlist=large-lun ostcount=4 \
mdsfilesystemtype=ldiskfs mdtfilesystemtype=ldiskfs \
ostfilesystemtype=ldiskfs
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I7207ce4c4e6771fc128c8b1d02d15d453b8556ae