LU-3561 tests: improve sanity-hsm.sh to support remote agent
In the current sanity-hsm.sh, only one HSM copytool is started
and the host of it (aka agent) is combined with the local Lustre
client node.

In order to test multiple remote HSM agents/copytools, this patch
introduces a new facet type named CPT and the following associated
variables:

- CPTCOUNT: number of copytools
- CPTDEV{N}: archive number/ID
- cpt{N}_HOST: hostname of the host where copytool cpt{N} is located
- SINGLECPT: facet of the single copytool

By default, the number of copytools are the number of remote client
nodes. Copytools will be started on the remote clients, not on the
local client. If there was no remote client, then the copytool will
be started on the local client.

Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I6e85d6948da894bd50148b0504347159ae23a79b