LU-3585 ptlrpc: Fix a crash when dereferencing NULL pointer
When a system runs out of memory and the function
ptlrpc_register_bulk() is called from ptl_send_rpc() the call to
LNetMEAttach() fails due to failure to allocate memory.  This forces
the code into an error path, which most probably previously went
untested.  The error path:
if (rc != 0) {
	CERROR("%s: LNetMEAttach failed x"LPU64"/%d: rc = %dn",
		desc->bd_export->exp_obd->obd_name, xid,
		posted_md, rc);
	break;
}
This print assumes that desc->bd_export is not NULL.  However, it is.
In fact it is expected to be NULL.  desc->bd_import is the correct
structure to access in this case.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I772ec757af224b9868956cc3fc9dc00e804fe84d