LU-3640 test: Sanity test 116a fails when an OST is full
Sanity test 116a looks for the OST with the least amount of free
space and fills it to varying levels. It does not check to see if
the OST is full and fails because of a divide by zero (bytes
available on full OST) error.

This patch adds a check to see if an OST is full and, if so, the test
is now skipped. We also corrected where to look for the qos_prio_free
file from the client to the MDS. A correction was made to how the
number of files written to each OST is calculated. We also added
using gos_threshold_rr to check how uneven the OSTs need to be.

This test is usually skipped since filling large OST takes too much
time. In order to get this test to run faster, we’ve added

1. A check to see if the OST needs to be filled to create uneven
OSTs. For QOS mode, a pair of OSTs must be at least qos_threshold_rr%
off balanced. If the OSTs are already at least uneven, we skip
writing files to an OST. If the OSTs do not exceed this threshold,
we write only 2% over the threshold amount to create uneven OSTs.

2. A cap on the number of files written when checking QOS
functionality. The files are small, 200k in size,and we were
writing enough files to fill 25% of the most full OST. Now we
cap the number of file written to at most 500. There’s no meaning
to 500 files, it just seemed like enough files to make sure that
the OST that has the most space available is getting more files.
We could make this number of files dependent on the number of OSTs.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ide1565015ffc26a849adabd0d06285971bcee9f2