LU-3601 ldlm: examine all waiting ibits locks on reprocess
In the reprocessing case of ldlm_process_inodebits_lock(), return
LDLM_ITER_CONTINUE even if the current lock could not be granted. This
will cause ldlm_reprocess_queue() to grant all locks that do not
conflict with a granted lock and do not conflict with a preceeding
(blocked) lock on the waiting list. This does not reorder conflicting
locks but does reorder the waiting list and makes the handling of
reprocess more similar to that of enqueue. The benefit this approach
is better liveness after lock cancellation (which may be required to
support HSM) at the expense making the worst case complexity of
reprocessing quadratic in the size of the waiting list.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: Ia854353fdbf63b2a52da1e95ce0f3c1ebe2f2682