LU-0000 osd-zfs: a proto for optimized object accounting
using dsl_sync_task_do_nowait() we can schedule a updates
to be executed in a closing txg. this way object accounting
can be done in memory and then updated in ZAPs once per txg.
this patch is just a proto for testing. it's still missing
important bits to export cached state to the Lustre quota
component (e.g. osd_acct_index_lookup())

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: Iebe96c9847c6c53b8232295a44fd8044b1ae18b4