LU-3647 hsm: do not verify data_version for restore
In ll_ioc_copy_end(), it used to verify data_version for restore.
This is not required because layout lock was already held by
restore request on the MDT side so that we can't verify whether
layout is stale. Therefore ll_data_version() uses stale layout to
call getattr to OFD which causes ENOENT, then this error code will
be transferred to MDT to result in restore failure.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: I8595d244399d84962d2af151b2d78459b10c56b4