LU-3666 tests: conf-sanity test_36 fix for remote osts
We removed the requirement that OSTs must be local to run
this tests. So, conf-sanity test_36 is not skipped when
OSTs are remote.
Since Lustre now allows for non 4k disk block size, create
OSTs with different blocksizes; 1k and 2k.
Fix ALLOWANCE to be calculated based on FSNAME2 ostcount,
instead of $OSTCOUNT.
Support added for ZFS.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Signed-off-by: Elena Gryaznova <elena_gryaznova@xyratex.com>
Xyratex-bug-id: MRP-896
Change-Id: I30295d015f9d81d26d31b25d059e2857d169bb3f