LU-3640 test: Sanity test 116a fails when an OST is full
Sanity test 116a looks for the OST with the least amount of
free space and fills it to varying levels. It does not check
to see if the OST is full and fails because of a divide by
zero (bytes avaiable on full OST).

A check was added to see if an OST is full and, if so, the test
is now skipped. We also corrected looking for the qos_prio_free
file on the client to now looking for it on the MDS.

We also need to see how long the test will take if we take
out the first check:
[ $MINV -gt 960000 ] && skip
because we skip this test over 95% of the time. It would
be nice to see if we still have to skip it if the OST
has a lot of free space.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ide1565015ffc26a849adabd0d06285971bcee9f2