LU-2825 lustre: Shrink sizes of cl_page and cl_req structures
For cl_page, changed the position of cp_ref, cp_state, cp_flags and
cp_type to better suit the alignment. Converted cp_flags and cp_type
members to bitfields. This leads to a saving of 16 bytes for cl_page
structure.

For cl_req, changed the position of crq_nrpages. This leads to a
saving of 8 bytes for cl_req structure.

Signed-off-by: Swapnil Pimpale <spimpale@ddn.com>
Change-Id: Id74e22cbed656f80f0fd8e63bf83c9dd680742e4