LU-3613 llapi: add call to get parent fids from link EA.
Implements llapi_path2parent() in liblustreapi.
This function returns the <parent_fid>/name(s) for a given entry
based on the link EA value.
Added '--parents' option to lfs path2fid to call this function.
Added sanity 164 to test it.

A part of this patch consists in removing useless dependencies
of obdclass/linkea.c on kernel code and liblustre symbols,
so it can be built for userland and linked with liblustreapi.
This required extracting lu_buf management out of lu_object.c,
and disabling kernel error injection macros when building llapi.

Adapt tests 39b and 162 to avoid cross-MDT links for current
DNE implementation.

Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Change-Id: I10f9316da605050be8fc709fe88d87abe86e22a4