LU-3326 test: sync after file deleted in run_dd.sh
in run_dd.sh, call "sync" after the file created by dd is deleted,
then the dirty pages of the deleted file will be flush to the OST
if MDT has not sent the object destroy request to it.

this patch also fixed the issue of missing LDLM_FL_DISCRAD_DATA
when sending blocking AST of LDLM extent lock.

Change-Id: Ia246122a5c3471b7cbea9cdaa7fad2ea4db856c5
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>