LU-3685 hsm: copy start error should set HP_FLAG_COMPLETED
If an error occurs when initializing a HSM request, in
ll_ioc_copy_start(), the PROGRESS message, sent to coordinator, should
carry the error code but also HP_FLAG_COMPLETED to mark the request as
finished (with error). If not, the Coordinator will ignore this
message and consider the request is still running.

Signed-off-by: Aurelien Degremont <aurelien.degremont@cea.fr>
Change-Id: Ie5ddbe8b47331202bd085941c681fcaea57ebf69