LU-3640 test: SANITY test 116a fails when an OST is full
Although a patch was submitted to see if an OST is full and,
if so, the test is now skipped, this doesn't fix the problem of
some test filling a single OST.

So this fortestonly patch is printing the space available
on each OST after various tests in order to find out
if the problem is a test leaving test files behind.

Filling min OST 50% to see if we can get the time of this
test down.

Also using this to see if there is an issue with
setstripe in test 27A.

Test-Parameters: fortestonly testlist=sanity,sanity
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I46a981ad17b140c0eb4869268cb440392b2b9de4