LU-3806 obdclass: add LCT_SERVER_SESSION for server session
Add LCT_SERVER_SESSION for server session, and separate the
server session flag from LCT_SESSION, so to avoid allocating
session info for client stack for each server request, if
client and server are on the same node.

This also prevents server crash when lustre module
is unloaded while IOs are in progress (LU-1165).

Signed-off-by: Wang Di <di.wang@intel.com>
Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: I68965374ce53d735186780d2d3f1bf083a0e8658