LU-2906 ldlm: use accessor macros for l_flags
Convert most all of the ldlm lock's l_flags references from direct
bit twiddling to using bit specific macros.  A few multi-bit
operations are left as an exercize for the reader.

The changes are mostly in ldlm, but also in llite, osc and quota,
but also fix a typo in the LDLM_{SET,CLEAR}_FLAG macros.

Signed-off-by: Bruce Korb <bruce_korb@xyratex.com>
Reviewed-by: Keith Mannthey <Keith.Mannthey@intel.com>
Xyratex-bug-id: MRP-783
Change-Id: I3291f088540b256af0a54a436f7ca96f064df7cc