LU-2906 ldlm: use accessor macros for l_flags
Convert most all of the ldlm lock's l_flags references from direct
bit twiddling to using bit specific macros.  A few multi-bit
operations are left as an exercise for the reader.

The changes are mostly in ldlm, but also in llite, osc and quota, but
also fix a typo in the LDLM_{SET,CLEAR}_FLAG macros and:

Add a multi-bit (mask) test.
Remove the now obsolete LDLM_AST_FLAGS and LDLM_INHERIT_FLAGS defines.
Remove the obsolete LDLM_FL_HIDE_LOCK_MASK define.
Rename "local_only" mask to "off_wire" since it is confusingly similar
to a flag that (I think) means, "do not copy this lock over the wire."
The "local_only/off_wire" mask is also never used.  It's mostly
informational.

Wireshark output moved to "lustre_dlm_flags_wshark.c" and only
bits that can actually appear "on the wire" are emitted.
The "packet-lustre.c" code that references these bits now gets
emitted into that file.  e.g. the "local_only" bit is never put
on the wire, so references to it in wireshark are gone.

Signed-off-by: Bruce Korb <bruce_korb@xyratex.com>
Reviewed-by: Keith Mannthey <Keith.Mannthey@intel.com>
Xyratex-bug-id: MRP-783
Change-Id: I3291f088540b256af0a54a436f7ca96f064df7cc