LU-136 build: add vmalloc option to allocate sbi->s_group_desc
Large kmalloc() for the group descriptor buffer pointers can
fail for large filesystems, which will cause the "not enough
memory" error while mounting. Fall back to vmalloc() if the
kmalloc() failed, as was done for s_flex_groups.

This patch is being revived to perform large-lun, 32TB,
testing on b1_8.

Signed-off-by: Yu Jian <yujian@whamcloud.com>
Change-Id: I09c4e51026825ec73fe30658bc062b277d2ff6e7
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: James Nunez <james.a.nunez@intel.com>