LU-3850 obdecho: create remote dir from echo client
Because echo client can only talk to local MDD, it will
send create remote dir req to the MDT where the dir is
located, which is different as the normal remote create
request. This patch fixes a few problem of this process

1. Not set parent during ah_init, which will make local OSD
to access the remote parent.

2. set correct attribute for remote parent in the declare
phase.

3. set root_fid for other MDT, which is needed by echo_client
to resolve the path on non-MDT0.

4. set child_base_id in oi_id directly to avoid the sanity
check in ostid_set_id,  and also child_base_id is actually
being used as the base of name, instead of FID.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: I16df8f0211afae670bfeea1ba1087a77a6b775da