LU-3872 tests: Fix rpmlint error in lustre/tests/Makefile.am
This patch is to fix the file permission of the file,
lustre/tests/clientapi/simple_test.c, in lustre/tests/Makefile.am,
which caused the rpmilnt check to fail with error:
"E: spurious-executable-perm (Badness: 50)".

The Makefile.am has simple_test.c as scripts (nobase_noinst_SCRIPTS),
so it is executable. It should be "data" (nobase_noinst_DATA) which
is installed 644.

Signed-off-by: Young Jong Pao <n16490@cray.com>
Change-Id: I83d692bccf8feba48b281a34636f6fd8ec2297f6