LU-3336 lfsck: recreate MDT-object or replace OST-object
If the MDT-object does not exist, but related OST-obejct still
alive, then when handle the orphan OST-object, the MDT-object
will be re-created under lost+found with the given OST-object
stripe information. The other parts of the MDT-obejct layout
EA will empty which will be filled by other orphan OST-object(s)
as the LFSCK processing.

If the MDT-object exists, but related layout EA slot is used
by some other OST-object stripe information, such OST-object
may be created by former LFSCK repairing for such MDT-object
which dangling reference. That means the former repairing is
based on local consistency guess, it maybe wrong. But later,
when the LFSCK find the real child OST-object, it will check
whether the former created OST-object for dangling reference
case has been modified or not. If not, destroy it, then fill
the layout EA slot with the real child OST-object.

Test-Parameters: testlist=sanity-lfsck
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ia7700f7b2e4e09d6b576d025295fdb5c10ca90e0