LU-4020 hsm: copytool event monitoring
Adds hooks into various llapi_hsm_* functions to emit JSON-formatted
event messages for consumption by a monitoring agent. The copytool
needs to be supplied with an optional --event-fifo argument to
enable this feature.

WIP: Pushing for proof of concept, but should be reasonably close
to production-quality, barring major concerns with design.

Signed-off-by: Michael MacDonald <michael.macdonald@intel.com>
Change-Id: I1fecfb8e60ef091c6d37d16dc1faf40be45c5ee2