LU-3884 lu: use proper flag for lu_object_put_nocache()
lu_object_put_nocache() was using dying flag
(LU_OBJECT_HEARD_BANSHEE) to indicate that the caller doesn't
want to cache current object, however, this introduces a
problem: if lu_object_put_nocache() is called for an object
which is holding by somebody else, the object will be marked
as dying but kept in cache until it released by the holder.

This patch introduced nocache flag (LU_OBJECT_NO_CACHE) for
lu_object_put_nocache(). The dying flag should only be used
for the object that is really dying soon.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: If82eb821b749c0c6d168992349fa79343d222bfc