LU-3867 utils: llapi_create_volatile_idx uses 0 instead idx
If llapi_create_volatile_idx() is passed an MDT index other than -1
then it uses '%.4X' to encode the index into the volatile file name.
But it uses an index of 0 rather than the one provided.

This patch removes the use of LUSTRE_VOLATILE_IDX and uses
LUSTRE_VOLATILE_HDR instead.

Signed-off-by: Bobbie Lind <bobbie.j.lind@intel.com>
Change-Id: I27e2beb8524683aba07c48297614dd02d7957af4